+++ tests/test_plotting.py
@@ -442,6 +442,18 @@ class TestSinglePlotter2D(unittest.TestCase):
         self.plotter.plot(dataset=test_dataset)
         self.assertEqual(cmap, self.plotter.drawing.cmap.name)
 
+    def test_plot_imshow_with_levels_ignores_levels(self):
+        self.plotter.parameters['levels'] = 40
+        test_dataset = dataset.Dataset()
+        test_dataset.data.data = np.random.random([5, 5])
+        test_dataset.plot(self.plotter)
+
+    def test_plot_imshow_sets_aspect_to_auto(self):
+        test_dataset = dataset.Dataset()
+        test_dataset.data.data = np.random.random([5, 5])
+        plotter = test_dataset.plot(self.plotter)
+        self.assertEqual('auto', self.plotter.ax._aspect)
+
 
 class TestSinglePlotter2DStacked(unittest.TestCase):
     def setUp(self):
@@ -940,13 +952,27 @@ class TestCompositePlotter(unittest.TestCase):
     def test_plot_sets_axes_position(self):
         self.plotter.grid_dimensions = [1, 1]
         self.plotter.subplot_locations = [[0, 0, 1, 1]]
-        self.plotter.axes_positions = [[0.3, 0.3, 0.6, 0.6]]
+        self.plotter.axes_positions = [[0.2, 0.2, -0.2, -0.2]]
         single_plotter = plotting.SinglePlotter1D()
         single_plotter.dataset = self.dataset
         self.plotter.plotter.append(single_plotter)
         self.plotter.plot()
-        self.assertEqual(self.plotter.axes_positions[0][0],
-                         list(self.plotter.axes[0].get_position().bounds)[0])
+        offsets = self.plotter.axes_positions[0]
+        axis_position = [0.125 + offsets[0]*0.775, 0.110 + offsets[1]*0.77,
+                         offsets[2]*0.775, offsets[3]*0.77]
+        self.assertListEqual(axis_position,
+                             list(self.plotter.axes[0].get_position().bounds))
+
+    def test_plot_shows_legend(self):
+        self.plotter.grid_dimensions = [1, 1]
+        self.plotter.subplot_locations = [[0, 0, 1, 1]]
+        single_plotter = plotting.SinglePlotter1D()
+        single_plotter.dataset = self.dataset
+        single_plotter.parameters['show_legend'] = True
+        self.plotter.plotter.append(single_plotter)
+        self.plotter.plot()
+        self.assertTrue(isinstance(self.plotter.axes[0].get_legend(),
+                                   matplotlib.legend.Legend))
 
 
 class TestSingleCompositePlotter(unittest.TestCase):