+++ docs/changelog.rst
@@ -5,6 +5,30 @@ Changelog
 This page contains a summary of changes between the official ASpecD releases. Only the biggest changes are listed here. A complete and detailed log of all changes is available through the `GitHub Repository Browser <https://github.com/tillbiskup/aspecd/commits/master>`_.
 
 
+Version 0.1.1
+=============
+
+Released 2021-05-03
+
+* MetadataMapper: Fix sequence of mapping operations performed
+
+* MetadataMapper: Mappings are automatically loaded from file if filename is given
+
+* CompositePlotter: Legends for subplots work
+
+* SliceExtraction: Remove correct axis from dataset
+
+* MultiPlotter1D*: Fix problem in conjunction with CompositePlotter and assigning drawings
+
+* SliceExtraction: fix problem extracting slice with index zero
+
+* CompositePlotter: more intuitive axes_positions (7b83f36c70)
+
+* Fix bug with aspect ratio of 2D plots using imshow
+
+* Update intersphinx mapping for matplotlib
+
+
 Version 0.1.0
 =============